# Improve and fix plugins loading