--- a/GitUI/CommandsDialogs/FormBrowse.cs
+++ b/GitUI/CommandsDialogs/FormBrowse.cs
@@ -30,6 +30,7 @@ using GitUI.UserControls.RevisionGrid;
 using GitUIPluginInterfaces;
 using GitUIPluginInterfaces.RepositoryHosts;
 using Microsoft;
+using Microsoft.VisualStudio.Threading;
 using Microsoft.Win32;
 using Microsoft.WindowsAPICodePack.Taskbar;
 using ResourceManager;
@@ -269,6 +270,8 @@ namespace GitUI.CommandsDialogs
 
             _formBrowseDiagnosticsReporter = new FormBrowseDiagnosticsReporter(this);
 
+            _repositoryHostsToolStripMenuItem.Visible = false;
+
             MainSplitContainer.Visible = false;
             MainSplitContainer.SplitterDistance = DpiUtil.Scale(260);
 
@@ -467,13 +470,53 @@ namespace GitUI.CommandsDialogs
 
             // All app init is done, make all repo related similar to switching repos
             SetGitModule(this, new GitModuleEventArgs(new GitModule(Module.WorkingDir)));
-
+            bool isDashboard = _dashboard?.Visible ?? false;
             ThreadHelper.FileAndForget(async () =>
             {
-                PluginRegistry.Initialize();
+                if (isDashboard)
+                {
+                    // Load only the git hoster plugin to quickly provide related features in dashboard
+                    PluginRegistry.InitializeGitHostersOnly();
+
+                    await this.SwitchToMainThreadAsync();
+
+                    // Now that hoster plugin is registered, populate Git-host-plugin actions on Dashboard, like "Clone GitHub repository"
+                    UpdateRepositoryHostsMenu();
+
+                    // Check if during plugin loading user left dashboard
+                    if (_dashboard.Visible)
+                    {
+                        _dashboard.RefreshContent();
+                    }
+
+                    // Come back to background to load all the other plugins (that could be long...)
+                    await TaskScheduler.Default;
+                }
+
+                await InitializeAndRegisterAllPluginsAsync();
+            });
+
+            return;
+
+            async Task InitializeAndRegisterAllPluginsAsync()
+            {
+                PluginRegistry.InitializeAll();
                 await this.SwitchToMainThreadAsync();
                 RegisterPlugins();
-            });
+                UpdateRepositoryHostsMenu();
+            }
+
+            void UpdateRepositoryHostsMenu()
+            {
+                if (PluginRegistry.GitHosters.Count != 0)
+                {
+                    // TODO: support more than one Git hosting provider plugin (ADO, GitLab,...)
+                    _repositoryHostsToolStripMenuItem.Text = PluginRegistry.GitHosters[0].Name;
+                }
+
+                // Show "Repository hosts" menu item when there is at least 1 repository host plugin loaded
+                _repositoryHostsToolStripMenuItem.Visible = PluginRegistry.GitHosters.Count != 0;
+            }
         }
 
         protected override void OnActivated(EventArgs e)
@@ -746,6 +789,11 @@ namespace GitUI.CommandsDialogs
         {
             foreach (ToolStripItem item in pluginsToolStripMenuItem.DropDownItems)
             {
+                if (item == pluginsLoadingToolStripMenuItem)
+                {
+                    continue;
+                }
+
                 item.Enabled = !(item.Tag is IGitPluginForRepository) || validWorkingDir;
             }
         }
@@ -757,6 +805,16 @@ namespace GitUI.CommandsDialogs
 
             lock (PluginRegistry.Plugins)
             {
+                if (PluginRegistry.Plugins.Count == 0)
+                {
+                    return;
+                }
+
+                if (pluginsToolStripMenuItem.DropDownItems.Contains(pluginsLoadingToolStripMenuItem))
+                {
+                    pluginsToolStripMenuItem.DropDownItems.Remove(pluginsLoadingToolStripMenuItem);
+                }
+
                 IOrderedEnumerable<IGitPlugin> pluginEntries = PluginRegistry.Plugins
                     .OrderByDescending(entry => entry.Name, StringComparer.CurrentCultureIgnoreCase);
 
@@ -793,14 +851,12 @@ namespace GitUI.CommandsDialogs
                     }
                     else
                     {
-                        pluginsToolStripMenuItem.DropDownItems.Insert(0, item);
-                    }
-                }
+                        // Handle special case where Git Hosting plugin has already been loaded
+                        ToolStripItem first = pluginsToolStripMenuItem.DropDownItems[0];
+                        int index = first is ToolStripSeparator || string.CompareOrdinal(first.Text, item.Text) >= 0 ? 0 : 1;
 
-                if (_dashboard?.Visible ?? false)
-                {
-                    // now that plugins are registered, populate Git-host-plugin actions on Dashboard, like "Clone GitHub repository"
-                    _dashboard.RefreshContent();
+                        pluginsToolStripMenuItem.DropDownItems.Insert(index, item);
+                    }
                 }
 
                 mainMenuStrip?.Refresh();
@@ -810,13 +866,6 @@ namespace GitUI.CommandsDialogs
             PluginRegistry.Register(UICommands);
 
             UICommands.RaisePostRegisterPlugin(this);
-
-            // Show "Repository hosts" menu item when there is at least 1 repository host plugin loaded
-            _repositoryHostsToolStripMenuItem.Visible = PluginRegistry.GitHosters.Count > 0;
-            if (PluginRegistry.GitHosters.Count == 1)
-            {
-                _repositoryHostsToolStripMenuItem.Text = PluginRegistry.GitHosters[0].Name;
-            }
 
             UpdatePluginMenu(Module.IsValidGitWorkingDir());
         }
@@ -835,7 +884,6 @@ namespace GitUI.CommandsDialogs
             pluginsToolStripMenuItem.Visible = false;
             refreshToolStripMenuItem.ShortcutKeys = Keys.None;
             refreshDashboardToolStripMenuItem.ShortcutKeys = Keys.None;
-            _repositoryHostsToolStripMenuItem.Visible = false;
             _formBrowseMenus.RemoveRevisionGridMainMenuItems();
             mainMenuStrip.Refresh();
         }