--- a/Plugins/GitUIPluginInterfaces/IGitPluginForCommit.cs
+++ b/Plugins/GitUIPluginInterfaces/IGitPluginForCommit.cs
@@ -1,0 +1,8 @@
+namespace GitUIPluginInterfaces;
+
+/// <summary>
+///  Plugin that will be loaded when FormCommit is opened
+/// </summary>
+public interface IGitPluginForCommit : IGitPluginForRepository
+{
+}