<!-- Please read CONTRIBUTING.md before submitting a pull request -->

Improve on "However, since the startup time of GitExtensions is so long, this feature is ruined for me." of https://github.com/gitextensions/gitextensions/issues/11518#issue-2081522202 (especially the first cold start)

## Proposed changes

- FormBrowse & Dashboard: Plugins start to be loaded a little later so that revision grid loading is not blocked anymore and is displayed more quickly (revision grid displayed around 4s earlier on a cold start)
- Dashboard: 
    - Load only the git hosters plugin (and later the others) making that "Clone GitHub repository" feature appears much more quickly (for that plugins need to export also "IRepositoryHostPlugin" interface)
    - fix state of GitHub menu items
- Form Commit **stand alone** (i.e. launched directly from file explorer): 
    - fix registering of plugins (that never worked)
    - improve performance by loading only "form commit" plugins (plugins need to export also new "IGitPluginForCommit" interface)

## Screenshots <!-- Remove this section if PR does not change UI -->

## Revision grid is not blocked by plugin loading

#### Before

Blank screen in the revision grid during 4 or 5s

![plugins_startup_before_slow](https://github.com/gitextensions/gitextensions/assets/460196/ea0597a2-417b-4119-b382-7f6051c16fd5)

#### After

Revision grid content appears immediately and the plugins are loaded after

Note: Here, to explain the principle, I was "lucky" because in this specific case the cold start was exceptionally long (around 15s to load the plugins) whereas normally it took around 5s. With current plugin loading, I would have waited more than 15s before seeing the revision grid loaded.
![plugins_startup_quick_best](https://github.com/gitextensions/gitextensions/assets/460196/d3ba98d4-3485-42e2-9bfa-d5b3cbdc7c12)

Same but with a usual time:

![plugins_startup_quick_normal](https://github.com/gitextensions/gitextensions/assets/460196/e9f57679-03c8-45db-aace-3a11b4565d62)

## Dashboard: Git provider loading

#### Before

![plugins_dashboard_slow](https://github.com/gitextensions/gitextensions/assets/460196/2016fda2-aea1-4bb6-8300-219c7c6bd894)

Note: I just discovered during the gif that it even slow down retrieving the checked out branches of all the repositories

#### After

![plugins_dashboard_quick](https://github.com/gitextensions/gitextensions/assets/460196/b9c84a5f-86c6-4cf9-bad7-051e44b1d656)


## Dashboard: fix state of GitHub menu items

#### Before

![image](https://github.com/gitextensions/gitextensions/assets/460196/337f388f-d4fd-4edc-9359-78d63dfaa0bc)

#### After

![image](https://github.com/gitextensions/gitextensions/assets/460196/9151f4c4-410f-4ac9-a576-9ee7855b4e5b)

## Test methodology <!-- How did you ensure quality? -->

- Manual

## Test environment(s) <!-- Remove any that don't apply -->

- Git Extensions 33.33.33
- Build 00049fdc9fc67ad0749232d84c886cbf9fb982fe
- Git 2.42.0.windows.2
- Microsoft Windows NT 10.0.22621.0
- .NET 8.0.0
- DPI 96dpi (no scaling)
- Portable: False
- Microsoft.WindowsDesktop.App Versions

```
    Microsoft.WindowsDesktop.App 6.0.21 [C:\Program Files\dotnet\shared\Microsoft.WindowsDesktop.App]
    Microsoft.WindowsDesktop.App 6.0.25 [C:\Program Files\dotnet\shared\Microsoft.WindowsDesktop.App]
    Microsoft.WindowsDesktop.App 7.0.14 [C:\Program Files\dotnet\shared\Microsoft.WindowsDesktop.App]
    Microsoft.WindowsDesktop.App 8.0.0 [C:\Program Files\dotnet\shared\Microsoft.WindowsDesktop.App]
```

## Merge strategy

<!-- Change the following if the merge strategy should be changed:
- Squash merge (maintainer to decide merge message, PR submitter should cleanup commits/messages at PR approval).
- Rebase merge (PR submitter must change the commit message for the last commit).
- Merge commit. (PR submitter to rebase and squash before merges).
- To be decided later.
The maintainer may still request the contributor to squash and rebase, to make sure that merges and commit messages are clarified.
-->

I agree that the maintainer **rebase** merge this PR (if the commit message is clear).

----

:black_nib: I contribute this code under [The Developer Certificate of Origin](../blob/master/contributors.txt).