@pmiossec 2024-01-23
Updates screenshots.