@gerhardol 2024-01-23
The first time starting GE is slow, it can take some seconds to start the app.
After that it is faster.
Debug builds are slower (I believe there are more files to check before loading there.)
I have for instance not been able to see the Plugins loading in Browse and load GH repo appears almost instantly in Dashboard.

This is an improvement for sure, I do not see anything quickly browsing the change.

Loading branch names in Dashboard is async, but I believe the mouse pointer is not changed.