@pmiossec 2024-02-12
@mstv I squashed all the fixup commits related to your previous review because it was becoming a mess.

But I kept the refactoring of the plugins registration new commit separate because I made some changes that simplify this part and I would like you to review it.